ref #929 created a PersonDbRepository. Added @Qualifier annotation to PersonRepository

+++ src/main/java/switch2019/project/domain/repositories/LedgerRepository.java
@@ -29,7 +29,7 @@ public interface LedgerRepository extends Repository {
      * @return
      */
 
-    int repositorySize();
+    long repositorySize();
 
     /**
      * Add a new ledger to Ledger Repository
