+++ src/main/java/switch2019/project/domain/repositories/Repository.java
@@ -34,7 +34,7 @@ public interface Repository {
      * @return
      */
 
-    int repositorySize();
+    long repositorySize();
 
 }
